+++ WinPaletter/Dialogs/ServiceInstaller.cs
@@ -31,7 +31,7 @@ namespace WinPaletter
             ApplyStyle(this);
             CheckForIllegalCrossThreadCalls = false;
 
-            Color c = PictureBox1.Image.AverageColor().CB((float)(Program.Style.DarkMode ? -0.35d : 0.35d));
+            Color c = PictureBox1.Image.AverageColor().CB(Program.Style.DarkMode ? -0.35f : 0.35f);
             AnimatedBox1.Color1 = c;
             AnimatedBox1.Color2 = Program.Style.Schemes.Main.Colors.AccentAlt;
